FIX: Added permission check to api_delete_alternative_file() [q10386]
git-svn-id: http://svn.resourcespace.com/svn/rs/trunk@11829 c08608d7-6e46-0410-86ca-f2a6f1370df5

+++ include/api_bindings.php
@@ -397,6 +397,11 @@ function api_add_alternative_file($resource, $name, $description = '', $file_nam
 
 function api_delete_alternative_file($resource,$ref)
 	{
+    global $disable_alternative_files;
+    if($disable_alternative_files || (0 < $resource && (!(get_edit_access($resource) || checkperm('A')))))
+        {
+        return false;
+        }
 	return delete_alternative_file($resource,$ref);
     }
